+++ NEWS.md
@@ -82,6 +82,7 @@ array of booleans that identifies the valid data.
 - `plot_strain_lm.py` no longer crashes when `window_args` is not provided.
 - `HierarchicalGridData` now owns the components.
 - Clear `OneGridFunction` cache in `grid_var` to avoid death by OOM.
+- Uniform constructor of `GridSeries` with constructors of other `Series`.
 
 #### New examples